#5: Move UnexpectedSequenceNumber to rangedberror package
