+++ store.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"log"
 	"strings"
 )
 
@@ -113,34 +112,6 @@ func ReadNRecords(totalEvents uint64, f func(context.Context) <-chan *Record) []
 	done()
 
 	return records
-}
-
-type UnexpectedSequenceNumber struct {
-	Expected           uint64
-	NextSequenceNumber uint64
-}
-
-// TODO: make private? convert to regex?
-func NewUnexpectedSequenceNumberFromString(input string) *UnexpectedSequenceNumber {
-	pieces := strings.Split(input, "unexpected sequence number:")
-	var expected, next uint64
-	_, err := fmt.Sscanf(pieces[1], "%d, next: %d", &expected, &next)
-	if err != nil {
-		log.Print(err)
-		return nil
-	}
-
-	return &UnexpectedSequenceNumber{
-		Expected:           expected,
-		NextSequenceNumber: next,
-	}
-}
-
-func (e UnexpectedSequenceNumber) Error() string {
-	return fmt.Sprintf("unexpected sequence number: %d, next: %d",
-		e.Expected,
-		e.NextSequenceNumber,
-	)
 }
 
 type rawEvent struct {
